requirements.txt
# Основные зависимости
requests>=2.28.0
psutil>=5.9.0

# Для скриншотов и изображений
Pillow>=9.0.0

# Для веб-камеры (опционально)
opencv-python>=4.6.0

# Для Linux систем
netifaces>=0.11.0

# Для обфускации (опционально)
pyarmor>=7.7.0

# Для шифрования (обязательно для reverse shell)
cryptography>=3.4.0

# Для сжатия (опционально)
zlib>=1.2.0
